Renamed rule-semicolon rules. Fix multi|single-line bugs. Closes #204.
