+++ src/rules/declaration-block-semicolon-newline-after/__tests__/index.js
