+++ src/rules/declaration-block-semicolon-newline-before/README.md
@@ -1,23 +1,14 @@
-# rule-semicolon-newline-after
+# declaration-block-semicolon-newline-before
 
-Require or disallow a newline after the semicolons of rules.
-
-```css
-    a {
-      color: pink;
-      top: 0;    ↑
-    }            ↑
-/**              ↑  
- * The newline after this semicolon */
-```
-
-End-of-line comments are allowed one space after the semicolon.
+Require or disallow a newline before the semicolons of rules.
 
 ```css
-a {
-  color: pink; /* something to say */
-  top: 0;
-}
+    a { 
+      color: pink 
+      ; top: 0;
+    } ↑
+/**   ↑  
+ * The newline before this semicolon */
 ```
 
 ## Options
@@ -26,38 +17,49 @@ a {
 
 ### `"always"`
 
-There *must always* be a single newline after the semicolon.
+There *must always* be a single newline before the semicolon.
 
 The following patterns are considered warnings:
 
 ```css
-a { color: pink; top: 0; }
+a { color: pink; }
+```
+
+```css
+a { 
+  color: pink; top: 0; 
+}
 ```
 
 The following patterns are *not* considered warnings:
 
 ```css
-a {
-  color: pink;
-  top: 0;
+a { color: pink
+; }
+```
+
+```css
+a { 
+  color: pink
+  ; top: 0; 
 }
 ```
 
 ### `"never"`
 
-There *must never* be whitespace after the semicolon.
+There *must never* be whitespace before the semicolon.
 
 The following patterns are considered warnings:
 
 ```css
-a { color: pink;
-}
+a { color: pink
+; }
 ```
 
 ```css
-a {
-  color: pink;
-  top: 0;
+a { 
+  color: pink
+  ; top: 0; 
 }
 ```
 
@@ -68,20 +70,20 @@ a { color: pink; }
 ```
 
 ```css
-a {
-  color: pink; top: 0;
+a { 
+  color: pink; top: 0; 
 }
 ```
 
 ### `"always-multi-line"`
 
-There *must always* be a single newline after the semicolon in multi-line rules.
+There *must always* be a single newline before the semicolon in multi-line rules.
 
 The following patterns are considered warnings:
 
 ```css
-a {
-  color: pink; top: 0;
+a { 
+  color: pink; top: 0; 
 }
 ```
 
@@ -96,22 +98,22 @@ a { color: pink; top: 0; }
 ```
 
 ```css
-a {
-  color: pink;
-  top: 0;
+a { 
+  color: pink
+  ; top: 0; 
 }
 ```
 
 ### `"never-multi-line"`
 
-There *must never* be whitespace after the semicolon in multi-line rules.
+There *must never* be whitespace before the semicolon in multi-line rules.
 
 The following patterns are considered warnings:
 
 ```css
-a {
-  color: pink;
-  top: 0;
+a { 
+  color: pink
+  ; top: 0; 
 }
 ```
 
@@ -126,8 +128,8 @@ a { color: pink; top: 0; }
 ```
 
 ```css
-a {
-  color: pink
-  ; top: 0;
+a { 
+  color: pink; 
+  top: 0; 
 }
 ```
