+++ src/rules/declaration-block-semicolon-newline-before/__tests__/index.js
