+++ src/rules/declaration-block-semicolon-space-after/__tests__/index.js
@@ -34,11 +34,13 @@ testRule("always-single-line", tr => {
   tr.ok("a { color: pink; }")
   tr.ok("a::before { content: \";a\"; }")
   tr.ok("a { color: pink; top: 0;}", "no space between trailing semicolon and closing brace")
+  tr.ok("a,\nb { color: pink; top: 0; }", "multi-line rule, single-line declaration-block")
 
   // Ignore multi-line
   tr.ok("a {\n  color: pink;\n  top: 0;\n}")
 
   tr.notOk("a { color: pink;top: 0; }", messages.expectedAfterSingleLine())
+  tr.notOk("a,\nb { color: pink;top: 0; }", messages.expectedAfterSingleLine())
   tr.notOk("a { color: pink;  top: 0; }", messages.expectedAfterSingleLine())
   tr.notOk("a { color: pink;\ttop: 0; }", messages.expectedAfterSingleLine())
 })
@@ -48,11 +50,13 @@ testRule("never-single-line", tr => {
   tr.ok("a { color: pink; }")
   tr.ok("a::before { content: \"; a\"; }")
   tr.ok("a { color: pink;top: 0; }", "space between trailing semicolon and closing brace")
+  tr.ok("a,\nb { color: pink;top: 0; }", "multi-line rule, single-line declaration-block")
 
   // Ignore multi-line
   tr.ok("a {\n  color: pink; top: 0;\n}")
 
   tr.notOk("a { color: pink; top: 0; }", messages.rejectedAfterSingleLine())
+  tr.notOk("a,\nb { color: pink; top: 0; }", messages.rejectedAfterSingleLine())
   tr.notOk("a { color: pink;  top: 0; }", messages.rejectedAfterSingleLine())
   tr.notOk("a { color: pink;\ttop: 0; }", messages.rejectedAfterSingleLine())
 })
